requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
boto3==1.26.104
botocore==1.29.165
cachecontrol==0.13.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==5.1.1
Django==4.1.7
django-admin-tailwind==0.0.31
django-cors-headers==3.14.0
django-crontab==0.7.1
django-recurrence==1.11.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
executing==1.2.0
fcm-django==2.0.0
firebase-admin==5.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
idna==3.4
jedi==0.19.0
jmespath==1.0.1
matplotlib-inline==0.1.6
msgpack==1.0.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.2
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6